refactor : moved report data to `models/`
